Fix more (MSVC) compiler warnings
reg. `size_t` vs. `int`, still not all (and some of these changes result in new warnings elsewhere), but 496 less than before (still 452 left).

+++ src/converter/internal/compat/notationmeta.cpp
@@ -60,8 +60,8 @@ mu::RetVal<std::string> NotationMeta::metaJson(notation::INotationPtr notation)
     json["poet"] =  poet(score);
     json["mscoreVersion"] =  score->mscoreVersion();
     json["fileVersion"] =  score->mscVersion();
-    json["pages"] =  score->npages();
-    json["measures"] =  score->nmeasures();
+    json["pages"] =  static_cast<int>(score->npages()); // no = operator for size_t
+    json["measures"] = static_cast<int>(score->nmeasures()); // no = operator for size_t
     json["hasLyrics"] =  boolToString(score->hasLyrics());
     json["hasHarmonies"] =  boolToString(score->hasHarmonies());
     json["keysig"] =  score->keysig();
